insights: refactor insight setting resolver out of telemetry package and into internal package for reuse in other code insights modules
